+++ src/neuroconv/datainterfaces/ecephys/neuroscope/neuroscopedatainterface.py
@@ -2,31 +2,17 @@
 from pathlib import Path
 from typing import Optional
 
-import spikeextractors as se
-from spikeinterface.core.old_api_utils import OldToNewRecording
 from pynwb.ecephys import ElectricalSeries
 
-from spikeinterface import BaseRecording
-from spikeinterface.extractors import NeuroScopeRecordingExtractor, NeuroScopeSortingExtractor
-
+from .neuroscope_utils import get_xml_file_path, get_channel_groups, get_shank_channels, get_session_start_time
 from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
 from ..baselfpextractorinterface import BaseLFPExtractorInterface
 from ..basesortingextractorinterface import BaseSortingExtractorInterface
+from ....tools import get_package
 from ....utils import FilePathType, FolderPathType, OptionalFilePathType, get_schema_from_hdmf_class, dict_deep_update
-from .neuroscope_utils import get_session_start_time
-
-try:
-    import lxml
-    from .neuroscope_utils import get_xml_file_path, get_channel_groups, get_shank_channels
-
-    HAVE_LXML = True
-
-except ImportError:
-    HAVE_LXML = False
-INSTALL_MESSAGE = "Please install lxml to use this interface!"
 
 
-def subset_shank_channels(recording_extractor: BaseRecording, xml_file_path: str) -> BaseRecording:
+def subset_shank_channels(recording_extractor, xml_file_path: str):
     """Attempt to create a SubRecordingExtractor containing only channels related to neural data."""
     shank_channels = get_shank_channels(xml_file_path=xml_file_path)
 
@@ -40,9 +26,7 @@ def subset_shank_channels(recording_extractor: BaseRecording, xml_file_path: str
     return sub_recording
 
 
-def add_recording_extractor_properties(
-    recording_extractor: BaseRecording, xml_file_path: str, gain: Optional[float] = None
-) -> BaseRecording:
+def add_recording_extractor_properties(recording_extractor, xml_file_path: str, gain: Optional[float] = None):
     """Automatically add properties to RecordingExtractor object."""
 
     if gain:
@@ -73,7 +57,7 @@ def add_recording_extractor_properties(
 class NeuroscopeRecordingInterface(BaseRecordingExtractorInterface):
     """Primary data interface class for converting a NeuroscopeRecordingExtractor."""
 
-    RX = NeuroScopeRecordingExtractor
+    ExtractorName = "NeuroScopeRecordingExtractor"
 
     @staticmethod
     def get_ecephys_metadata(xml_file_path: str):
@@ -120,13 +104,16 @@ class NeuroscopeRecordingInterface(BaseRecordingExtractorInterface):
             False by default. When True the interface uses the old extractor from the spikextractors library instead
             of a new spikeinterface object.
         """
-        assert HAVE_LXML, INSTALL_MESSAGE
+        get_package(package_name="lxml")
 
         if xml_file_path is None:
             xml_file_path = get_xml_file_path(data_file_path=file_path)
 
         if spikeextractors_backend:
-            self.RX = se.NeuroscopeRecordingExtractor
+            from spikeextractors import NeuroscopeRecordingExtractor
+            from spikeinterface.core.old_api_utils import OldToNewRecording
+
+            self.Extractor = NeuroscopeRecordingExtractor
             super().__init__(file_path=file_path, xml_file_path=xml_file_path, verbose=verbose)
             self.recording_extractor = OldToNewRecording(oldapi_recording_extractor=self.recording_extractor)
         else:
@@ -165,7 +152,8 @@ class NeuroscopeRecordingInterface(BaseRecordingExtractorInterface):
 class NeuroscopeMultiRecordingTimeInterface(NeuroscopeRecordingInterface):
     """Primary data interface class for converting a NeuroscopeMultiRecordingTimeExtractor."""
 
-    RX = se.NeuroscopeMultiRecordingTimeExtractor
+    RXModule = "spikeextractors"
+    RXName = "NeuroscopeMultiRecordingTimeExtractor"
 
     def __init__(
         self,
@@ -192,7 +180,8 @@ class NeuroscopeMultiRecordingTimeInterface(NeuroscopeRecordingInterface):
             If unspecified, it will be automatically set as the only .xml file in the same folder as the .dat file.
             The default is None.
         """
-        assert HAVE_LXML, INSTALL_MESSAGE
+        get_package(package_name="lxml")
+        from spikeinterface.core.old_api_utils import OldToNewRecording
 
         if xml_file_path is None:
             xml_file_path = get_xml_file_path(data_file_path=folder_path)
@@ -212,7 +201,7 @@ class NeuroscopeMultiRecordingTimeInterface(NeuroscopeRecordingInterface):
 class NeuroscopeLFPInterface(BaseLFPExtractorInterface):
     """Primary data interface class for converting Neuroscope LFP data."""
 
-    RX = NeuroScopeRecordingExtractor
+    ExtractorName = "NeuroScopeRecordingExtractor"
 
     def __init__(
         self,
@@ -240,13 +229,16 @@ class NeuroscopeLFPInterface(BaseLFPExtractorInterface):
             False by default. When True the interface uses the old extractor from the spikextractors library instead
             of a new spikeinterface object.
         """
-        assert HAVE_LXML, INSTALL_MESSAGE
+        get_package(package_name="lxml")
 
         if xml_file_path is None:
             xml_file_path = get_xml_file_path(data_file_path=file_path)
 
         if spikeextractors_backend:
-            self.RX = se.NeuroscopeRecordingExtractor
+            from spikeextractors import NeuroscopeRecordingExtractor
+            from spikeinterface.core.old_api_utils import OldToNewRecording
+
+            self.Extractor = NeuroscopeRecordingExtractor
             super().__init__(file_path=file_path, xml_file_path=xml_file_path)
             self.recording_extractor = OldToNewRecording(oldapi_recording_extractor=self.recording_extractor)
         else:
@@ -272,7 +264,7 @@ class NeuroscopeLFPInterface(BaseLFPExtractorInterface):
 class NeuroscopeSortingInterface(BaseSortingExtractorInterface):
     """Primary data interface class for converting a NeuroscopeSortingExtractor."""
 
-    SX = NeuroScopeSortingExtractor
+    ExtractorName = "NeuroScopeSortingExtractor"
 
     def __init__(
         self,
@@ -320,9 +312,11 @@ class NeuroscopeSortingInterface(BaseSortingExtractorInterface):
             False by default. When True the interface uses the old extractor from the spikextractors library instead
             of a new spikeinterface object.
         """
-        assert HAVE_LXML, INSTALL_MESSAGE
+        get_package(package_name="lxml")
+        from spikeextractors import NeuroscopeMultiSortingExtractor
+
         if spikeextractors_backend:
-            self.SX = se.NeuroscopeMultiSortingExtractor
+            self.Extractor = NeuroscopeMultiSortingExtractor
 
         super().__init__(
             folder_path=folder_path,