+++ src/neuroconv/datainterfaces/ophys/basesegmentationextractorinterface.py
@@ -1,27 +1,23 @@
 """Authors: Heberto Mayorquin, Cody Baker and Ben Dichter."""
-from abc import ABC
 from typing import Optional
 
 from pynwb import NWBFile
 from pynwb.device import Device
 from pynwb.ophys import Fluorescence, ImageSegmentation, ImagingPlane, TwoPhotonSeries
 
-from ...basedatainterface import BaseDataInterface
+from ...baseextractorinterface import BaseExtractorInterface
 from ...tools.roiextractors import write_segmentation, get_nwb_segmentation_metadata
-from ...utils import (
-    get_schema_from_hdmf_class,
-    fill_defaults,
-    OptionalFilePathType,
-    get_base_schema,
-)
+from ...utils import get_schema_from_hdmf_class, fill_defaults, OptionalFilePathType, get_base_schema
 
 
-class BaseSegmentationExtractorInterface(BaseDataInterface, ABC):
-    SegX = None
+class BaseSegmentationExtractorInterface(BaseExtractorInterface):
+    """Parent class for all SegmentationExtractorInterfaces."""
+
+    ExtractorModuleName: Optional[str] = "roiextractors"
 
     def __init__(self, **source_data):
         super().__init__(**source_data)
-        self.segmentation_extractor = self.SegX(**source_data)
+        self.segmentation_extractor = self.Extractor(**source_data)
 
     def get_metadata_schema(self):
         metadata_schema = super().get_metadata_schema()