+++ src/neuroconv/datainterfaces/ophys/extract/extractdatainterface.py
@@ -1,14 +1,9 @@
-from roiextractors import ExtractSegmentationExtractor
-from roiextractors.extraction_tools import FloatType
-
 from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
-from ....utils import FilePathType
+from ....utils import FilePathType, FloatType
 
 
 class ExtractSegmentationInterface(BaseSegmentationExtractorInterface):
     """Data interface for ExtractSegmentationExtractor."""
-
-    SegX = ExtractSegmentationExtractor
 
     def __init__(self, file_path: FilePathType, sampling_frequency: FloatType, verbose: bool = True):
         self.verbose = verbose