+++ src/neuroconv/tools/__init__.py
@@ -1,3 +1,7 @@
+# To be treated as sub-modules or plugins
 from .spikeinterface import spikeinterface
 from .roiextractors import roiextractors
 from .neo import neo
+
+# Basically utils but rely on external dependencies
+from .importing import get_package