+++ tests/test_minimal/test_schemas.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from neuroconv import interface_list
+from neuroconv.datainterfaces import interface_list
 from neuroconv.utils import load_dict_from_file