+++ tests/test_spikeinterface/test_interfaces.py
@@ -13,8 +13,8 @@ from spikeextractors.testing import check_recordings_equal, check_sortings_equal
 from hdmf.testing import TestCase
 from pynwb import NWBHDF5IO
 
-from neuroconv import NWBConverter, CEDRecordingInterface
-from neuroconv import SIPickleRecordingExtractorInterface, SIPickleSortingExtractorInterface
+from neuroconv import NWBConverter
+from neuroconv.datainterfaces import SIPickleRecordingInterface, SIPickleSortingInterface, CEDRecordingInterface
 from neuroconv.datainterfaces.ecephys.basesortingextractorinterface import BaseSortingExtractorInterface
 
 
@@ -25,8 +25,8 @@ from neuroconv.datainterfaces.ecephys.basesortingextractorinterface import BaseS
 class TestAssertions(TestCase):
     def test_import_assertions(self):
         with self.assertRaisesWith(
-            exc_type=AssertionError,
-            exc_msg="The sonpy package (CED dependency) is not available on Mac for Python versions below 3.8!",
+            exc_type=ModuleNotFoundError,
+            exc_msg="\nThe package 'sonpy' is not available on the darwin platform for Python version 3.7!",
         ):
             CEDRecordingInterface.get_all_channels_info(file_path="does_not_matter.smrx")
 
@@ -41,9 +41,7 @@ def test_pkl_interface():
     se.save_si_object(object_name="test_sorting", si_object=toy_data[1], output_folder=output_folder)
 
     class SpikeInterfaceTestNWBConverter(NWBConverter):
-        data_interface_classes = dict(
-            Recording=SIPickleRecordingExtractorInterface, Sorting=SIPickleSortingExtractorInterface
-        )
+        data_interface_classes = dict(Recording=SIPickleRecordingInterface, Sorting=SIPickleSortingInterface)
 
     source_data = dict(
         Recording=dict(file_path=str(test_dir / "test_pkl" / "test_recording.pkl")),
@@ -72,6 +70,8 @@ class TestSortingInterface(unittest.TestCase):
         sorting.add_unit(unit_id=3, times=np.arange(self.sorting_start_frames[2], self.num_frames))
 
         class TestSortingInterface(BaseSortingExtractorInterface):
+            Extractor = se.NumpySortingExtractor
+
             def __init__(self, verbose: bool = True):
                 self.sorting_extractor = sorting
                 self.source_data = dict()